+++ bridge-adapters/kinetic-bridgehub-adapter-ars-rest/src/main/java/com/kineticdata/bridgehub/adapter/ars/rest/ArsRestAdapter.java
@@ -12,6 +12,7 @@ import com.kineticdata.commons.v1.config.ConfigurablePropertyMap;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -28,6 +29,28 @@ import org.slf4j.LoggerFactory;
 
 public class ArsRestAdapter implements BridgeAdapter {
     /*----------------------------------------------------------------------------------------------
+     * CONSTRUCTOR
+     *--------------------------------------------------------------------------------------------*/
+    public ArsRestAdapter () {
+        // Parse the query and exchange out any parameters with their parameter 
+        // values. ie. change the query username=<%=parameter["Username"]%> to
+        // username=test.user where parameter["Username"]=test.user
+        this.parser = new ArsRestQualificationParser();
+    }
+    
+    /*----------------------------------------------------------------------------------------------
+     * STRUCTURES
+     *      AdapterMapping( Structure Name, Path Function)
+     *--------------------------------------------------------------------------------------------*/
+    public static Map<String,AdapterMapping> MAPPINGS 
+        = new HashMap<String,AdapterMapping>() {{
+        put("Entry", new AdapterMapping("Entry",
+            ArsRestAdapter::pathEntry));
+        put("Adhoc", new AdapterMapping("Adhoc",
+            ArsRestAdapter::pathAdhoc));
+    }};
+    
+    /*----------------------------------------------------------------------------------------------
      * PROPERTIES
      *--------------------------------------------------------------------------------------------*/
 
@@ -35,7 +58,7 @@ public class ArsRestAdapter implements BridgeAdapter {
     public static final String NAME = "Ars Rest Bridge";
 
     /** Defines the logger */
-    protected static final Logger logger = LoggerFactory.getLogger(ArsRestAdapter.class);
+    protected static final Logger LOGGER = LoggerFactory.getLogger(ArsRestAdapter.class);
     
     /** Adapter version constant. */
     public static String VERSION = "";
@@ -46,7 +69,7 @@ public class ArsRestAdapter implements BridgeAdapter {
             properties.load(ArsRestAdapter.class.getResourceAsStream("/"+ArsRestAdapter.class.getName()+".version"));
             VERSION = properties.getProperty("version");
         } catch (IOException e) {
-            logger.warn("Unable to load "+ArsRestAdapter.class.getName()+" version properties.", e);
+            LOGGER.warn("Unable to load "+ArsRestAdapter.class.getName()+" version properties.", e);
             VERSION = "Unknown";
         }
     }
@@ -71,7 +94,10 @@ public class ArsRestAdapter implements BridgeAdapter {
     private String password;
     private String origin;
     private ArsRestQualificationParser parser;
-    private ArsRestApiHelper arsApiHelper;
+    private ArsRestApiHelper apiHelper;
+    
+    // constant variables
+    private final String API_PATH = "/api/arsys/v1";
     
     /*---------------------------------------------------------------------------------------------
      * SETUP METHODS
@@ -85,10 +111,9 @@ public class ArsRestAdapter implements BridgeAdapter {
         password = properties.getValue(Properties.PROPERTY_PASSWORD);
         origin = properties.getValue(Properties.PROPERTY_ORIGIN);
         
-        arsApiHelper = new ArsRestApiHelper(origin, username, password);
-        parser = new ArsRestQualificationParser();
+        apiHelper = new ArsRestApiHelper(origin, username, password);
         
-        arsApiHelper.getToken();
+        apiHelper.getToken();
     }
 
     @Override
@@ -115,36 +140,32 @@ public class ArsRestAdapter implements BridgeAdapter {
         return properties;
     }
 
-    // Structures that are valid to use in the bridge. Used to check against
-    // when a method is called to make sure that the Structure the user is
-    // attempting to call is valid.
-    public static final List<String> VALID_STRUCTURES = Arrays.asList(new String[] {
-        "Entry"
-    });
-
     /*---------------------------------------------------------------------------------------------
      * IMPLEMENTATION METHODS
      *-------------------------------------------------------------------------------------------*/
 
     @Override
     public Count count(BridgeRequest request) throws BridgeError {
-        request.setQuery(substituteQueryParameters(request));
-        
         // Log the access
-        logger.trace("Counting records");
-        logger.trace("  Structure: " + request.getStructure());
-        logger.trace("  Query: " + request.getQuery());
+        LOGGER.trace("Counting records");
+        LOGGER.trace("  Structure: " + request.getStructure());
+        LOGGER.trace("  Query: " + request.getQuery());
 
-        // Check if the inputted structure is valid
-        if (!VALID_STRUCTURES.contains(request.getStructure())) {
-            throw new BridgeError("Invalid Structure: '" + 
-                request.getStructure() + "' is not a valid structure");
-        }
-
-        Map<String, String> parameters = parser.getParameters(request.getQuery());
+        // parse Structure
+        List<String> structureList = Arrays.asList(request.getStructure().trim()
+            .split("\\s*>\\s*"));
+        // get Structure model
+        AdapterMapping mapping = getMapping(structureList.get(0));
+        
+        Map<String, String> parameters = getParameters(
+            parser.parse(request.getQuery(),request.getParameters()), mapping);
+        
+        // Path builder functions may mutate the parameters Map;
+        String path = mapping.getPathbuilder().apply(structureList, parameters);
         
         // Retrieve the objects based on the structure from the source
-        JSONObject object = arsApiHelper.executeRequest(getUrl(request, parameters));
+        JSONObject object = apiHelper.executeRequest(getUrl(path, 
+            parameters));
         
         // Get domain specific data.
         JSONArray entries = (JSONArray)object.get("entries");
@@ -165,30 +186,36 @@ public class ArsRestAdapter implements BridgeAdapter {
 
     @Override
     public Record retrieve(BridgeRequest request) throws BridgeError {
-        request.setQuery(substituteQueryParameters(request));
-        
         // Log the access
-        logger.trace("Retrieving Kinetic Request CE Record");
-        logger.trace("  Structure: " + request.getStructure());
-        logger.trace("  Query: " + request.getQuery());
-        logger.trace("  Fields: " + request.getFieldString());
+        LOGGER.trace("Retrieving Kinetic Request CE Record");
+        LOGGER.trace("  Structure: " + request.getStructure());
+        LOGGER.trace("  Query: " + request.getQuery());
+        LOGGER.trace("  Fields: " + request.getFieldString());
 
-        // Check if the inputted structure is valid
-        if (!VALID_STRUCTURES.contains(request.getStructure())) {
-            throw new BridgeError("Invalid Structure: '" + request.getStructure() + "' is not a valid structure");
-        }
-
-        Map<String, String> parameters = parser.getParameters(request.getQuery());
+        // parse Structure
+        List<String> structureList = Arrays.asList(request.getStructure().trim()
+            .split("\\s*>\\s*"));
+        // get Structure model
+        AdapterMapping mapping = getMapping(structureList.get(0));
+        
+        Map<String, String> parameters = getParameters(
+            parser.parse(request.getQuery(),request.getParameters()), mapping);
+        
+        // Path builder functions may mutate the parameters Map;
+        String path = mapping.getPathbuilder().apply(structureList, parameters);
         
         // Retrieve the objects based on the structure from the source
-        JSONObject object = arsApiHelper.executeRequest(getUrl(request, parameters));
+        JSONObject object = apiHelper.executeRequest(getUrl(path, parameters));
         
         List<String> fields = request.getFields();
         if (fields == null) {
             fields = new ArrayList();
         }
         
-        // Get domain specific data.
+        // Get domain specific data.  If "multiple" were requested than obj will
+        // be null and entries will get populated.
+        // TODO: consider using mapper for single/multiple similar to kinetic
+        // core
         JSONObject obj = (JSONObject)(object).get("values");
         JSONArray entries = (JSONArray)object.get("entries");
         
@@ -198,12 +225,10 @@ public class ArsRestAdapter implements BridgeAdapter {
             Set<Object> removeKeySet = buildKeySet(fields, obj);
             obj.keySet().removeAll(removeKeySet);
 
-            // Create a Record object from the responce JSONObject
+            // Create a Record object from the JSONObject with removed keys
             record = new Record(obj);
         } else if (entries != null) {
             // Throw error is multiple results found.
-            // TODO: consider using mapper for single/multiple similar to kinetic
-            // core
             if (entries.size() > 1) {
                 throw new BridgeError ("Retrieve must return a single result."
                     + " Multiple results found.");
@@ -225,29 +250,25 @@ public class ArsRestAdapter implements BridgeAdapter {
 
     @Override
     public RecordList search(BridgeRequest request) throws BridgeError {
-        request.setQuery(substituteQueryParameters(request));
-        
         // Log the access
-        logger.trace("Searching Records");
-        logger.trace("  Structure: " + request.getStructure());
-        logger.trace("  Query: " + request.getQuery());
-        logger.trace("  Fields: " + request.getFieldString());
+        LOGGER.trace("Searching Records");
+        LOGGER.trace("  Structure: " + request.getStructure());
+        LOGGER.trace("  Query: " + request.getQuery());
+        LOGGER.trace("  Fields: " + request.getFieldString());
+        
+        // parse Structure
+        List<String> structureList = Arrays.asList(request.getStructure().trim()
+            .split("\\s*>\\s*"));
+        // get Structure model
+        AdapterMapping mapping = getMapping(structureList.get(0));
 
-        // Check if the inputted structure is valid
-        if (!VALID_STRUCTURES.contains(request.getStructure())) {
-            throw new BridgeError("Invalid Structure: '" + request.getStructure() + "' is not a valid structure");
-        }
-
-        Map<String, String> parameters = parser.getParameters(request.getQuery());
+        Map<String, String> parameters = getParameters(
+            parser.parse(request.getQuery(),request.getParameters()), mapping);
+        addLimit(parameters);
+        
         Map<String, String> metadata = request.getMetadata() != null ?
                 request.getMetadata() : new HashMap<>();
-     
-        addLimit(parameters);
-        // Add a sorting order to be used with the request if order was defined,
-        // but sort was not included in the qualification mapping.        
-        if (metadata.get("order") != null && !parameters.containsKey("sort")) {
-            addSort(metadata.get("order"), parameters);
-        }
+
         // If offest exists in metadata add it to the parameters for use with 
         // reqeust.  
         if (metadata.get("offset") != null) {
@@ -256,9 +277,18 @@ public class ArsRestAdapter implements BridgeAdapter {
         }
         // clear metadata object to be repopulated for response.
         metadata.clear();
+
+        // Add a sorting order to be used with the request if order was defined,
+        // but sort was not included in the qualification mapping.        
+        if (metadata.get("order") != null && !parameters.containsKey("sort")) {
+            addSort(metadata.get("order"), parameters);
+        }
+
+        // Path builder functions may mutate the parameters Map;
+        String path = mapping.getPathbuilder().apply(structureList, parameters);
         
         // Retrieve the objects based on the structure from the source
-        JSONObject object = arsApiHelper.executeRequest(getUrl(request, parameters));
+        JSONObject object = apiHelper.executeRequest(getUrl(path, parameters));
 
         // Get domain specific data.
         JSONArray entries = (JSONArray)object.get("entries");
@@ -280,7 +310,7 @@ public class ArsRestAdapter implements BridgeAdapter {
             // Set object to user defined fields
             Set<Object> removeKeySet = buildKeySet(fields, entry);
 
-            // Iterate through the responce objects and make a new Record for each.
+            // Iterate through the response objects and make a new Record for each.
             for (Object o : entries) {
                 entry = (JSONObject)((JSONObject)o).get("values");
 
@@ -322,7 +352,7 @@ public class ArsRestAdapter implements BridgeAdapter {
             metadata.put("offset", Integer.toString((limit + 1) + offset));
         
         } catch (NumberFormatException e) {
-            logger.error("Error parsing int: " + e);
+            LOGGER.error("Error parsing int: ", e);
         }
     }
     
@@ -335,7 +365,7 @@ public class ArsRestAdapter implements BridgeAdapter {
                 limit = Integer.parseInt(parameters.get("limit").trim());
                 if (limit < 0 || limit > 1000) {
                     limit = 1000;
-                    logger.debug("limit was outside standard values. Limit set "
+                    LOGGER.debug("limit was outside standard values. Limit set "
                         + "to 1000 default.");
                 }
                 parameters.replace("limit", Integer.toString(limit));
@@ -343,38 +373,13 @@ public class ArsRestAdapter implements BridgeAdapter {
                 parameters.put("limit", "1000");
             }
         } catch (NumberFormatException e) {
-            logger.error("limit parmaeter must be a number.  limit set to 1000 "
-                + "default. " + e);
+            LOGGER.error("limit parmaeter must be a number.  limit set to 1000 "
+                + "default. ", e);
         }
     }
     
-    // Take the sort order from metadata and add it to parameters for use with
-    // request.  Encoding field names and joining fields with comma is required
-    // due to ARS 9 api behavior.  Encoded commas between field namesbreaks api 
-    // requests. 
-    protected void addSort(String order, 
-        Map<String, String> parameters) throws BridgeError {
-        
-        LinkedHashMap<String,String> sortOrderItems = getSortOrderItems(
-                BridgeUtils.parseOrder(order));
-            String str = sortOrderItems.entrySet().stream().map(entry -> {
-                String key = "";
-                try {
-                    key = URLEncoder.encode(entry.getKey().trim(), "UTF-8");
-                }   catch (UnsupportedEncodingException e) {
-                    logger.error("Error encoding sort order for field: " 
-                        + entry.getKey() + " " + e);
-                    return "";
-                }
-                return key + "." + entry.getValue().toLowerCase();
-
-
-            }).collect(Collectors.joining(","));
-            parameters.put("sort", str);
-    }
-    
     // Create a set of keys to remove from object prior to creating Record.
-    // TODO: consider using fields(foo,bar) to only request required feilds from
+    // TODO: consider using fields(foo,bar) to only request required fields from
     // api.  This would eliminate the need to manipulate return objects
     protected Set<Object> buildKeySet(List<String> fields, JSONObject obj) {
         if(fields.isEmpty()){
@@ -388,33 +393,11 @@ public class ArsRestAdapter implements BridgeAdapter {
             if(fields.contains(key)){
                 continue;
             }else{
-                logger.trace("Remove Key: "+key);
+                LOGGER.trace("Remove Key: "+key);
                 removeKeySet.add(key);
             }
         }
         return removeKeySet;
-    }
-    
-    // Build url for request.  Encode all parameters except the sort parameter.
-    // The encoding for sort is done in the addSort method.  Read comment for 
-    // explanation.
-    protected String getUrl (BridgeRequest request, Map<String, String> parameters) {
-                
-        String str = parameters.entrySet().stream().map(entry -> {
-            if (!entry.getKey().equals("sort")) {
-                try {
-                    return entry.getKey() + "=" 
-                        + URLEncoder.encode(entry.getValue(), "UTF-8");
-                } catch (UnsupportedEncodingException e) {
-                    logger.error("Error encoding query parameter: " + e);
-                }
-                return entry.getKey() + "=" + entry.getValue();
-            }
-            return entry.getKey() + "=" + entry.getValue();
-        }).collect(Collectors.joining("&"));
-        
-        return String.format("%s/api/arsys/v1/%s?%s", origin, 
-            parser.parsePath(request.getQuery()), str);
     }
     
     private LinkedHashMap<String, String> 
@@ -430,12 +413,183 @@ public class ArsRestAdapter implements BridgeAdapter {
         
         return (LinkedHashMap)uncastSortOrderItems;
     }
+        
+    protected JSONArray getResponseData(Object responseData) {
+        JSONArray responseArray = new JSONArray();
+        
+        if (responseData instanceof JSONArray) {
+            responseArray = (JSONArray)responseData;
+        }
+        else if (responseData instanceof JSONObject) {
+            // It's an object
+            responseArray.add((JSONObject)responseData);
+        }
+        
+        return responseArray;
+    }
     
-    private String substituteQueryParameters(BridgeRequest request) throws BridgeError {
-        // Parse the query and exchange out any parameters with their parameter 
-        // values. ie. change the query username=<%=parameter["Username"]%> to
-        // username=test.user where parameter["Username"]=test.user
-        ArsRestQualificationParser parser = new ArsRestQualificationParser();
-        return parser.parse(request.getQuery(),request.getParameters());
+    /**
+     * This helper is intended to abstract the parser get parameters from the core
+     * methods.
+     * 
+     * @param request
+     * @param mapping
+     * @return
+     * @throws BridgeError 
+     */
+    protected Map<String, String> getParameters(String query,  
+        AdapterMapping mapping) throws BridgeError {
+        
+        Map<String, String> parameters = new HashMap<>();
+        if (mapping.getStructure() == "Adhoc") {
+            // Adhoc qualifications are two segments. ie path?queryParameters
+            String [] segments = query.split("[?]",2);
+
+            // getParameters only needs the queryParameters segment
+            if (segments.length > 1) {
+                parameters = parser.getParameters(segments[1]);
+            }
+            // Pass the path along to the functional operator
+            parameters.put("adapterPath", segments[0]);
+        } else {
+            parameters = parser.getParameters(query);
+        }
+        
+        return parameters;
+    }
+        
+    /**
+     * This method checks that the structure on the request matches on in the 
+     * Mapping internal class.  Mappings map directly to the adapters supported 
+     * Structures.  
+     * 
+     * @param structure
+     * @return Mapping
+     * @throws BridgeError 
+     */
+    protected AdapterMapping getMapping (String structure) throws BridgeError{
+        AdapterMapping mapping = MAPPINGS.get(structure);
+        if (mapping == null) {
+            throw new BridgeError("Invalid Structure: '" 
+                + structure + "' is not a valid structure");
+        }
+        return mapping;
+    }
+    
+    /**
+     * Build url for request.  Encode all parameters except the sort parameter.
+     * The encoding for sort is done in the addSort method.  Read comment for 
+     * explanation.
+     * 
+     * @param path
+     * @param parameters
+     * @return String
+     */
+    protected String getUrl (String path, Map<String, String> parameters) {
+                
+        String str = parameters.entrySet().stream().map(entry -> {
+            if (!entry.getKey().equals("sort")) {
+                try {
+                    return entry.getKey() + "=" 
+                        + URLEncoder.encode(entry.getValue(), "UTF-8");
+                } catch (UnsupportedEncodingException e) {
+                    LOGGER.error("Error encoding query parameter: " + e);
+                }
+                return entry.getKey() + "=" + entry.getValue();
+            }
+            return entry.getKey() + "=" + entry.getValue();
+        }).collect(Collectors.joining("&"));
+        
+        return String.format("%s%s?%s", API_PATH, path, str);
+    }
+ 
+    /**
+     * Take the sort order from metadata and add it to parameters for use with
+     * request.  Encoding field names and joining fields with comma is required
+     * due to ARS 9 api behavior.  Encoded commas between field names breaks api 
+     * requests. 
+     * 
+     * @param order
+     * @param parameters
+     * @return
+     * @throws BridgeError 
+     */
+    protected void addSort(String order, 
+        Map<String, String> parameters) throws BridgeError {
+        
+        LinkedHashMap<String,String> sortOrderItems = getSortOrderItems(
+            BridgeUtils.parseOrder(order));
+        String str = sortOrderItems.entrySet().stream().map(entry -> {
+            String key = "";
+            try {
+                key = URLEncoder.encode(entry.getKey().trim(), "UTF-8");
+            }   catch (UnsupportedEncodingException e) {
+                LOGGER.error("Error encoding sort order for field: " 
+                    + entry.getKey() + " ", e);
+                return "";
+            }
+            return key + "." + entry.getValue().toLowerCase();
+
+        }).collect(Collectors.joining(","));
+        
+        parameters.put("sort", str);
+    }
+    
+    /**************************** Path Definitions ****************************/
+    /**
+     * Build the path for the Deals structure.
+     * 
+     * @param structureList
+     * @param parameters
+     * @return 
+     * @throws com.kineticdata.bridgehub.adapter.BridgeError 
+     */
+    protected static String pathEntry(List<String> structureList,
+        Map<String, String> parameters) throws BridgeError {
+
+        if (!(structureList.size() > 1)) {
+            throw new BridgeError("The Entry structure requires a Form Name.");
+        }
+        
+        String path = String.format("%s/%s","/entry", structureList.get(1));
+        if (parameters.containsKey("entry_id")) {
+            path = String.format("%s/%s", path, parameters.get("entry_id"));
+            parameters.remove("entry_id");
+        }
+
+        return path;
+    }
+    
+    /**
+     * Build path for Adhoc structure.
+     * 
+     * @param structureList
+     * @param parameters
+     * @return
+     * @throws BridgeError 
+     */
+    protected static String pathAdhoc(List<String> structureList, 
+        Map<String, String> parameters) throws BridgeError {
+        
+        return parameters.get("adapterPath");
+    }
+
+    /**
+     * Checks if a parameter exists in the parameters Map.
+     * 
+     * @param param
+     * @param parameters
+     * @param structureList
+     * @throws BridgeError 
+     */
+    protected static void checkRequiredParamForStruct(String param,
+        Map<String, String> parameters, List<String> structureList)
+        throws BridgeError{
+        
+        if (!parameters.containsKey(param)) {
+            String structure = String.join(" > ", structureList);
+            throw new BridgeError(String.format("The %s structure requires %s"
+                + "parameter.", structure, param));
+        }
     }
 }