+++ bridge-adapters/kinetic-bridgehub-adapter-ars-rest/src/test/java/com/kineticdata/bridgehub/adapter/ars/rest/ArsRestTest.java
@@ -1,16 +1,13 @@
 package com.kineticdata.bridgehub.adapter.ars.rest;
 
-import com.kineticdata.bridgehub.adapter.BridgeAdapter;
 import com.kineticdata.bridgehub.adapter.BridgeAdapterTestBase;
 import com.kineticdata.bridgehub.adapter.BridgeError;
 import com.kineticdata.bridgehub.adapter.BridgeRequest;
 import com.kineticdata.bridgehub.adapter.Count;
 import com.kineticdata.bridgehub.adapter.Record;
 import com.kineticdata.bridgehub.adapter.RecordList;
-import com.kineticdata.bridgehub.adapter.ars.rest.ArsRestAdapter;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import org.junit.Test;
@@ -40,13 +37,13 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("id");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People?q='Remedy Login ID'=\"<%=parameter[\"Id\"]%>\"");
+        request.setQuery("q='Remedy Login ID'=\"<%=parameter[\"Id\"]%>\"");
         
-        Map parameters = new HashMap();
-        parameters.put("Id", "kinetic_integration_user");
-        request.setParameters(parameters);
+        request.setParameters(new HashMap<String, String>() {{ 
+            put("Id", "test_user");
+        }});
         
         Count count = null;
         try {
@@ -65,16 +62,16 @@ public class ArsRestTest extends BridgeAdapterTestBase{
 
         // Create the Bridge Request
         List<String> fields = new ArrayList<String>();
-        fields.add("id");
+        fields.add("id-x");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People?q='Remedy Login sID'=\"<%=parameter[\"Id\"]%>\"");
+        request.setQuery("q='Remedy Login sID'=\"<%=parameter[\"Id\"]%>\"");
         
-        Map parameters = new HashMap();
-        parameters.put("Id", "_tiggebz");
-        request.setParameters(parameters);
+        request.setParameters(new HashMap<String, String>() {{ 
+            put("Id", "test_user");
+        }});
         
         error = null;
         try {
@@ -102,16 +99,20 @@ public class ArsRestTest extends BridgeAdapterTestBase{
     }
     
     @Test
-    public void test_retrieve() throws Exception{
+    public void test_retrieve_single() throws Exception{
         BridgeError error = null;
         
         // Create the Bridge Request
         List<String> fields = new ArrayList<String>();
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CST:CHSAttributes");
         request.setFields(fields);
-        request.setQuery("entry/CST:CHSAttributes/000000000003430/attach/Schematic");
+        request.setQuery("entry_id=<%=parameter[\"Id\"]%>");
+        
+        request.setParameters(new HashMap<String, String>() {{ 
+            put("Id", "000000000003430");
+        }});
         
         Record record = null;
         try {
@@ -125,7 +126,7 @@ public class ArsRestTest extends BridgeAdapterTestBase{
     }
     
     @Test
-    public void test_retrieve_single() throws Exception{
+    public void test_retrieve() throws Exception{
         BridgeError error = null;
         
         // Create the Bridge Request
@@ -134,12 +135,12 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("Last Modified Date");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People?q='Remedy Login ID'=\"<%=parameter[\"Login ID\"]%>\"");
+        request.setQuery("q='Remedy Login ID'=\"<%=parameter[\"Login ID\"]%>\"");
         
         Map parameters = new HashMap();
-        parameters.put("Login ID", "marcin.zarycki@kineticdata.com");
+        parameters.put("Login ID", "test_user");
         request.setParameters(parameters);
         
         Record record = null;
@@ -165,9 +166,65 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("Last Name");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People");
+        request.setQuery("");
+        
+        RecordList records = null;
+        try {
+            records = getAdapter().search(request);
+        } catch (BridgeError e) {
+            error = e;
+        }
+        
+        assertNull(error);
+        assertTrue(records.getRecords().size() > 0);
+    }
+    
+    @Test
+    public void test_search_adhoc() throws Exception{
+        BridgeError error = null;
+        
+        assertNull(error);
+        
+        // Create the Bridge Request
+        List<String> fields = new ArrayList<String>();
+        fields.add("First Name");
+        fields.add("Last Name");
+        
+        BridgeRequest request = new BridgeRequest();
+        request.setStructure("Adhoc");
+        request.setFields(fields);
+        request.setQuery("/entry/CTM:People");
+        
+        RecordList records = null;
+        try {
+            records = getAdapter().search(request);
+        } catch (BridgeError e) {
+            error = e;
+        }
+        
+        assertNull(error);
+        assertTrue(records.getRecords().size() > 0);
+    }
+        
+    @Test
+    public void test_search_adhoc_parameters() throws Exception{
+        BridgeError error = null;
+        
+        // Create the Bridge Request
+        List<String> fields = new ArrayList<String>();
+        fields.add("First Name");
+        fields.add("Last Name");
+        
+        BridgeRequest request = new BridgeRequest();
+        request.setStructure("Adhoc");
+        request.setFields(fields);
+        request.setQuery("/entry/CTM:People?q='Remedy Login ID'  LIKE \"<%=parameter[\"Login ID\"]%>\"&sort=Last+Name.desc,First+Name.desc");
+        
+        Map parameters = new HashMap();
+        parameters.put("Login ID", "_%");
+        request.setParameters(parameters);
         
         RecordList records = null;
         try {
@@ -190,9 +247,9 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("Last Name");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People?sort=Last+Name.desc,First+Name.desc");
+        request.setQuery("sort=Last+Name.desc,First+Name.desc");
         
         RecordList records = null;
         try {
@@ -215,9 +272,9 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("Last Name");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People");
+        request.setQuery("");
         
         Map <String, String> metadata = new HashMap<>();
         metadata.put("order", "<%=field[\"Last Name\"]%>:DESC");       
@@ -244,9 +301,9 @@ public class ArsRestTest extends BridgeAdapterTestBase{
         fields.add("Last Name");
         
         BridgeRequest request = new BridgeRequest();
-        request.setStructure("Entry");
+        request.setStructure("Entry > CTM:People");
         request.setFields(fields);
-        request.setQuery("entry/CTM:People");
+        request.setQuery("");
         
         Map <String, String> metadata = new HashMap<>();
         metadata.put("order", "<%=field[\"Last Name\"]%>:DESC"